Fix: ensure array-typed properties are arrays
Ensure Polymer element properties that are typed with Array are always
arrays.

The `microtesia.js` library only creates array properties when it
encounters multiple values with the same name in microdata, For example:

  <div itemscope>
    <div itemprop="names">val</div>
  </div>

Would result in `names = "val"` even though we might expect the `names`
property to be array-typed.

However, Polymer properties are typed and we were ending up with
non-array values being populated in array-typed properties as in the
above example.

This change fixes this by detecting when we have single valuex for
array-typed propertids and converting thm to a single-entry arrays
instead.